+++ src/components/MemoAddStep2.vue
@@ -15,28 +15,6 @@ export default {
   name: 'MemoAddStep2',
   computed: {
     ...mapGetters('jumpstart', ['get_to', 'get_subject', 'get_date', 'get_description'])
-  },
-  methods: {
-    ...mapActions('jumpstart', ['test']),
-    async submit() {
-      try {
-        await this.$store.dispatch('memo/test', {
-          params: {
-            body: {
-              date: this.date,
-              description: this.description,
-              to: this.to,
-              subject: this.subject
-            }
-          }
-        });
-        alert('Successfully sent');
-      } catch (error) {
-        alert('Not sent, an error occured.');
-
-        console.log(error);
-      }
-    }
   }
 };
 </script>
